Break Temporal runtime dependency in Workflow and Executor
